Updated migrations

+++ migrations/20230512153941_convert_dates_to_numbers.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
         tbl.dropColumn('date');
     })
     .table('property_events', tbl => {
-        tbl.integer('date').defaultTo(new Date().getTime());
+        tbl.bigInteger('date').defaultTo(new Date().getTime());
     })
 };
