+++ migrations/20230518154859_replace_energy_usage_with_usage.js
@@ -6,7 +6,7 @@ exports.up = function(knex) {
   return knex.schema.dropTableIfExists('energy_usage').createTable('usage', tbl => {
     tbl.string('type');
     tbl.float('price').defaultTo(0);
-    tbl.integer('time').defaultTo(new Date().getTime());
+    tbl.bigInteger('time').defaultTo(new Date().getTime());
     tbl.integer('property_id').notNullable();
   });
 };
@@ -18,7 +18,7 @@ exports.up = function(knex) {
 exports.down = function(knex) {
     return knex.schema.dropTableIfExists('usage').createTable('energy_usage', tbl => {
         tbl.float('price').defaultTo(0);
-        tbl.integer('time').defaultTo(new Date().getTime());
+        tbl.bigInteger('time').defaultTo(new Date().getTime());
         tbl.integer('property_id').notNullable();
     });
 };
